Final Commit

+++ EducationalInstitute/src/main/java/com/college/ed/model/Group.java
@@ -1,8 +1,6 @@
 package com.college.ed.model;
 
 import java.util.List;
-import java.util.Set;
-
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -12,27 +10,31 @@ import jakarta.persistence.Id;
 import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 
+// Creating group entity
 @Entity
 @Table(name="\"Group\"")
 public class Group {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @Column(name="groupName", columnDefinition = "VARCHAR(255)")
 	private String groupName;
 	
+    // One To Many relationship with routine entity
 	@OneToMany(fetch =FetchType.LAZY, mappedBy="group")
 	private List<Routine> routine;
     
+	// Default Constructors
     public Group() {
 	}
-
-	public Long getGroupId() {
+    
+    // Getters and Setters
+	public int getGroupId() {
 		return id;
 	}
 
-	public void setGroupId(Long id) {
+	public void setGroupId(int id) {
 		this.id = id;
 	}
 
@@ -44,6 +46,7 @@ public class Group {
 		this.groupName = groupName;
 	}
 
+	//String toString
 	@Override
 	public String toString() {
 		return "Group [id=" + id + ", groupName=" + groupName + ", getGroupId()=" + getGroupId()
