+++ EducationalInstitute/src/main/java/com/college/ed/repository/GroupRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.college.ed.model.Group;
 
-public interface GroupRepository extends JpaRepository<Group,Long>{
+public interface GroupRepository extends JpaRepository<Group,Integer>{
+
 
 }
