+++ EducationalInstitute/src/main/java/com/college/ed/repository/RoutineRepository.java
@@ -8,10 +8,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.college.ed.model.Routine;
 import com.college.ed.model.Teacher;
 
-public interface RoutineRepository extends JpaRepository<Routine,Long>{
+public interface RoutineRepository extends JpaRepository<Routine,Integer>{
 
 	List<Routine> findByTeacherAndRoutineDateBetween(Teacher teacher, LocalDate startDate, LocalDate endDate);
 
-	List<Routine> findByGroupId(Long groupId);
+	List<Routine> findByGroupId(int groupId);
 
 }
